Work on #74, repackaged and added javadocs
