--- reactor/src/main/java/com/iluwatar/reactor/ChannelHandler.java
@@ -1,8 +1,0 @@
-package com.iluwatar.reactor;
-
-import java.nio.channels.SelectionKey;
-
-public interface ChannelHandler {
-
-	void handleChannelRead(AbstractNioChannel channel, Object readObject, SelectionKey key);
-}
