--- reactor/src/main/java/com/iluwatar/reactor/Dispatcher.java
@@ -1,8 +1,0 @@
-package com.iluwatar.reactor;
-
-import java.nio.channels.SelectionKey;
-
-public interface Dispatcher {
-	void onChannelReadEvent(AbstractNioChannel channel, Object readObject, SelectionKey key);
-	void stop();
-}
