+++ reactor/src/test/java/com/iluwatar/reactor/app/AppTest.java
@@ -1,4 +1,4 @@
-package com.iluwatar.reactor;
+package com.iluwatar.reactor.app;
 
 import java.io.IOException;
